requirements.txt
numpy
pandas
python-dotenv ##helps in reading env file
mysql-connector-python
pymysql #helps in connecting database
scikit-learn